+++ RELEASES.md
@@ -24,6 +24,7 @@ Upcoming
 - Show http body in exception thrown by http sender in case status code indicates an error
 - Make multipart work for http sender in case only inputMessageParam is used (without extra parameters)
 - Bugfix RestListenerServlet that didn't read http body anymore for POST method
+- Add support for paramsInUrl, inputMessageParam and multipart to Larva HttpSender